Support end2end tests under CMake (for Vulkan). (#391)
Add support for building tests through CMakelist #386

+++ src/gpgmm/utils/Math.h
@@ -33,7 +33,7 @@ namespace gpgmm {
 
     bool IsPowerOfTwo(uint64_t number);
     uint64_t NextPowerOfTwo(uint64_t number);
-    bool IsAligned(uint32_t number, size_t multiple);
+    bool IsAligned(uint64_t number, size_t multiple);
 
     template <typename T>
     T AlignToPowerOfTwo(T number, size_t alignment) {
